+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+### Version 1.5.0
+__Changes__
+- added KButil_Remove_Genomes_from_GenomeSet()
+- changed KButil_Add_Genomes_to_GenomeSet() to accept genome object list
+- Description input field no longer required
+
 ### Version 1.4.0
 __Changes__
 - added KButil_Batch_Create_ReadsSet()